+++ android-pubsub/src/main/java/com/pj/pubsub/MessageMediator.kt
@@ -1,9 +1,0 @@
-package com.pj.pubsub
-
-import com.pj.pubsub.extensions.Message
-import com.pj.pubsub.proto.NativePubSub.Envelope
-
-interface MessageMediator {
-    fun register(node: ReceivablePublisher)
-    fun publish(envelope: Envelope, tag: Tag)
-}