+++ android-pubsub/src/main/java/com/pj/pubsub/PublisherType.kt
@@ -1,8 +1,0 @@
-package com.pj.pubsub
-
-enum class PublisherType(val defaultID: Int){
-    Android     (10000),
-    IOS         (20000),
-    Unity       (30000),
-    Unreal      (40000)
-}