+++ android-pubsub/src/main/java/com/pj/pubsub/unity/NativeBridgeCallback.java
@@ -1,5 +1,0 @@
-package com.pj.pubsub.unity;
-
-public interface NativeBridgeCallback{
-    void onReceive(byte[] data);
-}