+++ minipubsub/src/main/java/com/minisdk/pubsub/Node.kt
@@ -1,0 +1,10 @@
+package com.minisdk.pubsub
+
+import com.minisdk.pubsub.data.IdCounter
+
+abstract class Node{
+    companion object{
+        private val idCounter = IdCounter()
+    }
+    val id : Int = idCounter.getNext()
+}