+++ minipubsub/src/main/java/com/minisdk/pubsub/bridge/NativeBridgeCallback.java
@@ -1,0 +1,5 @@
+package com.minisdk.pubsub.bridge;
+
+public interface NativeBridgeCallback{
+    void onReceive(String info, String data);
+}