+++ minipubsub/src/main/java/com/minisdk/pubsub/data/Topic.kt
@@ -1,0 +1,11 @@
+package com.minisdk.pubsub.data
+
+import com.google.gson.annotations.JsonAdapter
+
+data class Topic(
+    val key: String,
+    @JsonAdapter(SdkTypeAdapter::class)
+    val target: SdkType
+)
+
+val defaultTopic = Topic("", SdkType.Native)