+++ sample/src/main/java/com/pj/sample/SampleKitLoader.java
@@ -1,0 +1,7 @@
+package com.pj.sample;
+
+public class SampleKitLoader{
+    public static void load() {
+        SampleKit.INSTANCE.prepare();
+    }
+}